refactor/#8 : dir sercurity
